Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
    "playwright>=1.40.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=14.2.0",
    "selenium>=4.15.0",
    "supabase>=2.3.0",
    "webdriver-manager>=4.0.0",
]
